workspace/requirements.txt
accelerate==1.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
diffusers==0.35.2
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
hf-xet==1.2.0
hf_transfer==0.1.9
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numba==0.62.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.1
psutil==7.1.3
PyMatting==1.1.14
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rembg==2.0.68
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
sympy==1.14.0
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
zipp==3.23.0
